requirements.txt
pytest==7.4.0
h11==0.14.0
pip==22.3.1
attrs==22.2.0
wheel==0.38.4
cryptography==41.0.3
requests==2.31.0
colorama==0.4.6
trio==0.22.1
outcome==1.2.0
sniffio==1.3.0
exceptiongroup==1.1.2
sortedcontainers==2.4.0
cffi==1.15.1
idna==3.4
mysql-connector-python==8.1.0
setuptools==65.5.1
Appium-Python-Client==1.3.0
selenium==4.1.0
urllib3==1.26.16
protobuf==4.21.12
pluggy==1.2.0
iniconfig==2.0.0
certifi==2023.5.7
six==1.16.0
pycparser==2.21
websocket-client==0.59.0
eyes-selenium==5.20.0
marshmallow==2.21.0
packaging==23.1
wsproto==1.2.0
tqdm==4.65.0
python-dotenv==1.0.0